Extract nftAccess smart loan tests to a separate file
